+++ src/components/Swapping.tsx
@@ -18,7 +18,7 @@ import { oneInch } from '../services/1Inch'
 import { AnySwapExecutionManager } from '../services/anyswap.execute'
 import { getBalancesForWallet } from '../services/balanceService'
 import { CbridgeExecutionManager } from '../services/cbridge.execute'
-import { HopExecutionManager } from '../services/hop_new.execute'
+import { HopExecutionManager } from '../services/hop.execute'
 import { HorizonExecutionManager } from '../services/horizon.execute'
 import { lifinance } from '../services/lifinance'
 import { storeActiveRoute } from '../services/localStorage'
