+++ src/services/hop.execute.ts
@@ -1,11 +1,13 @@
-/* eslint-disable max-params */
-import { JsonRpcSigner } from '@ethersproject/providers'
+import { TransactionResponse } from '@ethersproject/abstract-provider'
+import { constants } from 'ethers'
 
-import { CoinKey } from '../types'
-import { Execution } from '../types/'
+import { ExecuteCrossParams, getChainById } from '../types'
+import { checkAllowance } from './allowance.execute'
 import hop from './hop'
+import Lifi from './LIFI/Lifi'
 import notifications, { NotificationType } from './notifications'
 import { createAndPushProcess, initStatus, setStatusDone, setStatusFailed } from './status'
+import { personalizeStep } from './utils'
 
 export class HopExecutionManager {
   shouldContinue: boolean = true
@@ -14,86 +16,111 @@ export class HopExecutionManager {
     this.shouldContinue = val
   }
 
-  executeCross = async (
-    signer: JsonRpcSigner,
-    bridgeCoin: CoinKey,
-    amount: string,
-    fromChainId: number,
-    toChainId: number,
-    updateStatus?: Function,
-    initialStatus?: Execution,
-  ) => {
-    // setup
-    const { status, update } = initStatus(updateStatus, initialStatus)
-    hop.init(signer, fromChainId, toChainId)
+  executeCross = async ({ signer, step, updateStatus }: ExecuteCrossParams) => {
+    const { action, execution, estimate } = step
+    const { status, update } = initStatus(updateStatus, execution)
+    const fromChain = getChainById(action.fromChainId)
+    const toChain = getChainById(action.toChainId)
 
-    //set allowance AND send
-    const allowanceAndCrossProcess = createAndPushProcess(
-      'allowanceAndCrossProcess',
-      update,
-      status,
-      'Set Allowance and Cross',
-    )
-    let tx
-    try {
-      if (allowanceAndCrossProcess.txHash) {
-        tx = signer.provider.getTransaction(allowanceAndCrossProcess.txHash)
-      } else {
+    // STEP 1: Check Allowance ////////////////////////////////////////////////
+    // approval still needed?
+    const oldCrossProcess = status.process.find((p) => p.id === 'crossProcess')
+    if (!oldCrossProcess || !oldCrossProcess.txHash) {
+      if (action.fromToken.id !== constants.AddressZero) {
+        // Check Token Approval only if fromToken is not the native token => no approval needed in that case
         if (!this.shouldContinue) return status
-        tx = await hop.setAllowanceAndCrossChains(bridgeCoin, amount, fromChainId, toChainId)
-        allowanceAndCrossProcess.txHash = tx.hash
+        await checkAllowance(
+          signer,
+          fromChain,
+          action.fromToken,
+          action.fromAmount,
+          estimate.approvalAddress,
+          update,
+          status,
+          true,
+        )
+      }
+    }
+
+    // STEP 2: Get Transaction ////////////////////////////////////////////////
+    const crossProcess = createAndPushProcess('crossProcess', update, status, 'Prepare Transaction')
+
+    try {
+      let tx: TransactionResponse
+      if (crossProcess.txHash) {
+        // load exiting transaction
+        tx = await signer.provider.getTransaction(crossProcess.txHash)
+      } else {
+        // create new transaction
+        const personalizedStep = await personalizeStep(signer, step)
+        const { tx: transactionRequest } = await Lifi.getStepTransaction(personalizedStep)
+
+        // STEP 3: Send Transaction ///////////////////////////////////////////////
+        crossProcess.status = 'ACTION_REQUIRED'
+        crossProcess.message = 'Sign Transaction'
+        update(status)
+        if (!this.shouldContinue) return status
+
+        tx = await signer.sendTransaction(transactionRequest)
+
+        // STEP 4: Wait for Transaction ///////////////////////////////////////////
+        crossProcess.status = 'PENDING'
+        crossProcess.txHash = tx.hash
+        crossProcess.txLink = fromChain.metamask.blockExplorerUrls[0] + 'tx/' + crossProcess.txHash
+        crossProcess.message = 'Wait for'
         update(status)
       }
+
+      await tx.wait()
     } catch (e: any) {
-      if (e.message) allowanceAndCrossProcess.errorMessage = e.message
-      if (e.code) allowanceAndCrossProcess.errorCode = e.code
-      setStatusFailed(update, status, allowanceAndCrossProcess)
+      if (e.message) crossProcess.errorMessage = e.message
+      if (e.code) crossProcess.errorCode = e.code
+      setStatusFailed(update, status, crossProcess)
       throw e
     }
-    setStatusDone(update, status, allowanceAndCrossProcess)
 
-    //wait for transaction
-    if (!this.shouldContinue) return status
+    crossProcess.message = 'Transfer started: '
+    setStatusDone(update, status, crossProcess)
+
+    // STEP 5: Wait for Receiver //////////////////////////////////////
     const waitForTxProcess = createAndPushProcess(
       'waitForTxProcess',
       update,
       status,
-      'Wait for transaction on receiving chain',
+      'Wait for Receiving Chain',
     )
     let destinationTxReceipt
     try {
+      hop.init(signer, action.fromChainId, action.toChainId)
       destinationTxReceipt = await hop.waitForDestinationChainReceipt(
-        allowanceAndCrossProcess.txHash,
-        bridgeCoin,
-        fromChainId,
-        toChainId,
+        crossProcess.txHash,
+        action.toToken.key,
+        action.fromChainId,
+        action.toChainId,
       )
-      waitForTxProcess.destinationReceipt = destinationTxReceipt.transactionHash
-      update(status)
-    } catch (_e) {
-      const e = _e as Error
-      if (e.message)
-        waitForTxProcess.errorMessage = 'Transaction failed on receiving chain: \n' + e.message
+    } catch (e: any) {
+      waitForTxProcess.errorMessage = 'Failed waiting'
+      if (e.message) waitForTxProcess.errorMessage += ':\n' + e.message
+      if (e.code) waitForTxProcess.errorCode = e.code
       notifications.showNotification(NotificationType.CROSS_ERROR)
       setStatusFailed(update, status, waitForTxProcess)
       throw e
     }
 
-    const parsedReceipt = hop.parseReceipt(tx, destinationTxReceipt)
-
-    if (!this.shouldContinue) return status
-    setStatusDone(update, status, waitForTxProcess, {
-      fromAmount: parsedReceipt.fromAmount,
-      toAmount: parsedReceipt.toAmount,
-      gasUsed: parsedReceipt.gasUsed,
-    })
-
-    // -> set status
+    // -> parse receipt & set status
+    const parsedReceipt = hop.parseReceipt(crossProcess.txHash, destinationTxReceipt)
+    waitForTxProcess.txHash = destinationTxReceipt.transactionHash
+    waitForTxProcess.txLink =
+      toChain.metamask.blockExplorerUrls[0] + 'tx/' + waitForTxProcess.txHash
+    waitForTxProcess.message = 'Funds Received:'
+    status.fromAmount = parsedReceipt.fromAmount
+    status.toAmount = parsedReceipt.toAmount
+    // status.gasUsed = parsedReceipt.gasUsed
     status.status = 'DONE'
-    notifications.showNotification(NotificationType.CROSS_SUCCESSFUL)
-    update(status)
+    setStatusDone(update, status, waitForTxProcess)
 
     // DONE
+    notifications.showNotification(NotificationType.CROSS_SUCCESSFUL)
     return status
   }
 }
